backend/medschool/requirements.txt
asgiref==3.8.1
django==5.1
sqlparse==0.4.4
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1